+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 * Fixed a bug that prevented saving searches on the loadouts page
+* Improved Loadout Optimizer speed.
 
 ## 8.131.0 <span class="changelog-date">(2026-06-28)</span>