+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fixed a bug that prevented saving searches on the loadouts page
 * Improved Loadout Optimizer speed.
+* Actually show all items for sale in Eververse after Monument of Triumph release.
 
 ## 8.131.0 <span class="changelog-date">(2026-06-28)</span>