+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed a bug that prevented saving searches on the loadouts page
 * Improved Loadout Optimizer speed.
 * Actually show all items for sale in Eververse after Monument of Triumph release.
+* Only render community perk description links that are http(s) URLs
 
 ## 8.131.0 <span class="changelog-date">(2026-06-28)</span>